Made unwind handling cross-platform, with some limitation on Windows unwind info that needs Windows syscall for now (for #3508)
